pyproject.toml
[tool.poetry]
name = "esd-services-api-client"
version = "0.0.0"
description = "Python clients for ESD services"
authors = ["ECCO Sneaks & Data <esdsupport@ecco.com>"]
maintainers = ['GZU <gzu@ecco.com>', 'JRB <ext-jrb@ecco.com>', 'VISA <visa@ecco.com>']
readme = "README.md"
license = 'Apache 2.0'
repository = 'https://github.com/SneaksAndData/esd-services-api-client'

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
adapta = { version="^2.6.5", extras = ["azure", "storage", "datadog"] }
dataclasses-json = "^0.6.0"
pycryptodome = "~3.15"
azure-identity = { version = "~1.7", optional = true }
injector = { version = "~0.21.0", optional = true }
pyjwt = "~2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
pylint = "^2.12"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12"
requests = "~2.27"
cryptography = "~36.0"
requests-mock = "~1.10"
locust = "~2.14.2"
black = "^23"
responses = "~0.20"
pytest-timeout = "2.1.0"

[build-system]
requires = ["poetry-core", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
azure = [
    'azure-identity'
]

nexus = [
    'injector'
]
